docs: make WHY more clear
In fact, Rails STI is just perfect, doesn’t require any configuration and it works out of the box. Except when dealing with legacy databases or third-party systems that depend on the same database, we are out of luck. This is WHY this gem was born and WHY it will be useful to others and it could be part of Rails.

Most other web frameworks or ORMs would offer something like this.

- [Sequel single-table inheritance plugin](https://sequel.jeremyevans.net/rdoc-plugins/classes/Sequel/Plugins/SingleTableInheritance.html)
- [Django Proxy](https://dev.to/zachtylr21/model-inheritance-in-django-m0j)
- [Java JPA discriminator](https://openjpa.apache.org/builds/1.0.2/apache-openjpa-1.0.2/docs/manual/jpa_overview_mapping_discrim.html)